requirements.txt
numpy==2.1.3
pillow==11.1.0
streamlit==1.44.1
tensorflow==2.19.0

